refactor(search): restructure SearchBar components and enhance error handling
- Moved HighlightSearchResults to a shared directory for better organization.
- Updated SearchBar to handle loading errors and improved error logging.
- Enhanced the fetching logic for search documents and lunr index with error handling.
- Introduced new shared components and styles for better maintainability.
