+++ preview/src/theme/SearchBar/shared/HighlightSearchResults.jsx
