+++ preview/src/theme/SearchBar/shared/algolia.css
