+++ preview/src/theme/SearchBar/shared/styles.css
