+++ preview/src/theme/SearchBar/shared/templates.js
